implement product repository

+++ internal/domain/repository/dto/product.go
@@ -1,15 +1,15 @@
 package dto
 
 type CreateProduct struct {
-	Name     string `json:"name" validate:"required"`
-	Price    uint   `json:"price" validate:"required"`
-	Quantity uint   `json:"quantity" validate:"required"`
+	Name     string  `json:"name" validate:"required"`
+	Price    float64 `json:"price" validate:"required"`
+	Quantity int     `json:"quantity" validate:"required"`
 }
 
 type UpdateProduct struct {
-	Name     string `json:"name" validate:"required"`
-	Price    uint   `json:"price" validate:"required"`
-	Quantity uint   `json:"quantity" validate:"required"`
+	Name     string  `json:"name" validate:"required"`
+	Price    float64 `json:"price" validate:"required"`
+	Quantity int     `json:"quantity" validate:"required"`
 }
 
 type ProductQuery struct {
